+++ src/main/java/org/openlmis/stockmanagement/service/referencedata/ApprovedProductReferenceDataService.java
@@ -20,11 +20,12 @@ import static org.apache.commons.collections.CollectionUtils.isEmpty;
 import org.openlmis.stockmanagement.dto.referencedata.ApprovedProductDto;
 import org.openlmis.stockmanagement.dto.referencedata.OrderableDto;
 import org.openlmis.stockmanagement.util.RequestParameters;
+import org.openlmis.stockmanagement.web.Pagination;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+
 import java.util.Collection;
+import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -52,19 +53,13 @@ public class ApprovedProductReferenceDataService extends
    *
    * @param facilityId id of the facility
    * @param programId  id of the program
-   * @param pageable   pagination parameters
    * @return a collection of approved products matching the search criteria
    */
   public Page<OrderableDto> getApprovedProducts(UUID facilityId, UUID programId,
-                                                Collection<UUID> orderableIds, Pageable pageable) {
+                                                Collection<UUID> orderableIds) {
     RequestParameters params = RequestParameters.init();
 
     params.set("programId", programId);
-
-    if (null != pageable) {
-      params.set("size", pageable.getPageSize());
-      params.set("page", pageable.getPageNumber());
-    }
 
     if (!isEmpty(orderableIds)) {
       params.set("orderableId", orderableIds);
@@ -73,8 +68,12 @@ public class ApprovedProductReferenceDataService extends
     Page<ApprovedProductDto> approvedProductPage =
         getPage(facilityId + "/approvedProducts", params);
 
-    return new PageImpl<>(approvedProductPage.getContent().stream()
-        .map(ApprovedProductDto::getOrderable).collect(Collectors.toList()),
-        pageable, approvedProductPage.getTotalElements());
+    List<OrderableDto> content = approvedProductPage
+        .getContent()
+        .stream()
+        .map(ApprovedProductDto::getOrderable)
+        .collect(Collectors.toList());
+
+    return Pagination.getPage(content);
   }
 }