+++ src/test/java/org/openlmis/stockmanagement/service/referencedata/OrderableFulfillReferenceDataServiceTest.java
@@ -17,22 +17,31 @@ package org.openlmis.stockmanagement.service.referencedata;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.eq;
 import static org.powermock.api.mockito.PowerMockito.doReturn;
 import static org.powermock.api.mockito.PowerMockito.when;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.type.TypeFactory;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.openlmis.stockmanagement.dto.referencedata.OrderableFulfillDto;
+import org.openlmis.stockmanagement.service.AuthService;
 import org.openlmis.stockmanagement.testutils.OrderableFulfillDtoDataBuilder;
-import org.openlmis.stockmanagement.util.RequestParameters;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.test.util.ReflectionTestUtils;
+import org.springframework.web.client.RestTemplate;
+
+import java.net.URI;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -44,8 +53,13 @@ public class OrderableFulfillReferenceDataServiceTest {
   @Mock
   private ObjectMapper objectMapper;
 
+  @Mock
+  private AuthService authService;
+
+  @Mock
+  private RestTemplate restTemplate;
+
   @InjectMocks
-  @Spy
   private OrderableFulfillReferenceDataService service;
 
   private UUID id = UUID.randomUUID();
@@ -57,6 +71,8 @@ public class OrderableFulfillReferenceDataServiceTest {
     TypeFactory factory = mapper.getTypeFactory();
 
     when(objectMapper.getTypeFactory()).thenReturn(factory);
+    when(authService.obtainAccessToken()).thenReturn("token");
+    ReflectionTestUtils.setField(service, "restTemplate", restTemplate);
   }
 
   @Test
@@ -64,8 +80,10 @@ public class OrderableFulfillReferenceDataServiceTest {
     Map<UUID, OrderableFulfillDto> map = new HashMap<>();
     map.put(UUID.randomUUID(), new OrderableFulfillDtoDataBuilder().build());
 
-    doReturn(map)
-        .when(service).findOne(any(RequestParameters.class));
+    doReturn(new ResponseEntity<>(map, HttpStatus.OK))
+        .when(restTemplate)
+        .exchange(any(URI.class), eq(HttpMethod.GET), any(HttpEntity.class), eq(Map.class));
+
     when(objectMapper.convertValue(any(Object.class), any(JavaType.class)))
         .thenReturn(map);